♻️  refactor: onChange 함수들 따로 작성하지 않고 직접 작성
